Added a separate check command for library
The main problem this tries to solve is that in a complex project which
consists of a library and binaries which use it it's much more important
to iterate on the library itself rather than on binaries and `check`
command is pretty important.

While solving this issue in a general case might be complicated (it
needs binary name), the solution for library case is quite simple - all
we need to do is to pass `--lib`.

This commit adds additional internal command `rust.cargo.check.lib` and
exposes it as:
1. "Cargo: Check Library" task
2. `cargo-lib` as a new option for `checkWith` workspace configuration
   value

So the easiest way is to set `checkWith` to `cargo-lib`

Fixes #131