Skipping hidden lines in doc examples
As per the [Rust book](https://doc.rust-lang.org/book/documentation.html), lines in codeblocks in documentation comments starting with `#` are not rendered. I think they should be hidden in the docs hover too – code checks if line starts with `#` (and is a inside a code block) and skips this line.