Don't follow include! expansion to origin
include! errors are special in the sense that caller side is usually not useful for reporting all the syntax / type errors originating in included file (as you're getting tons of messages pointing to the same line where include!(...) occurs).

This change adds special case that keeps errors bound to their original location if it was include!d.