Fix rustfmt not working on Windows
rustfmt 0.9.0 uses canonicalized paths, which look like this on Windows:
`\\?\C:\Some\Path\file.rs`. These paths don't compare properly to the
.rs.fmt files that vscode-rust is working with. Through a ton of ugly Uri
hacks, I was able to turn them into paths that can be properly compared to
the ones in rustfmt's diffs. This is super ugly and might not even work on
macOS or Linux, so it needs some testing there.

Fixes #270